noop split resources.go -> resources{,_dev,_rp}.go
